Use classes for common client
